Move multiplicity counter
