+++ executor/src/witgen/data_structures/mod.rs
@@ -1,3 +1,4 @@
 pub mod column_map;
 pub mod copy_constraints;
 pub mod finalizable_data;
+pub mod multiplicity_counter;
