+++ executor/src/witgen/data_structures/multiplicity_counter.rs
